pyproject.toml
[project]
name = "awesome-rag-attacks"
version = "0.1.0"
description = "A collection of awesome RAG attacks"
authors = [
    { name = "Awesome RAG Attacks", email = "awesome-rag-attacks@example.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core LangChain
    "langchain>=0.2.14",
    "langchain-community>=0.2.14",
    "langchain-text-splitters>=0.2.14",
    
    # LLM Providers
    "langchain-openai>=0.2.14",
    "langchain-huggingface>=0.2.14",
    
    # Vector Store
    "faiss-cpu>=1.7.0",
    
    # Embeddings and ML
    "sentence-transformers",
    "torch>=2.0.0",
    "transformers",
    
    # Configuration and Utilities
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "loguru>=0.7.0",
    
    # Data Processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    
    # BEIR dataset support (for compatibility with PoisonedRAG)
    "beir>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]